+++ src/aiotools/utils.py
@@ -5,21 +5,17 @@ A set of higher-level coroutine aggregation utilities based on :class:`Superviso
 from __future__ import annotations
 
 import asyncio
-from contextlib import aclosing
-from contextvars import Context
-from typing import (
-    Any,
+from collections.abc import (
     AsyncGenerator,
     Awaitable,
     Coroutine,
     Generator,
     Iterable,
-    List,
-    Optional,
     Sequence,
-    Tuple,
-    TypeVar,
 )
+from contextlib import aclosing
+from contextvars import Context
+from typing import Any, Optional, TypeVar
 
 from .supervisor import Supervisor
 
@@ -104,7 +100,7 @@ async def gather_safe(
     coros: Iterable[Coroutine[Any, None, T] | Generator[None, None, T]],
     *,
     context: Optional[Context] = None,
-) -> List[T | Exception]:
+) -> list[T | BaseException]:
     """
     A safer version of :func:`asyncio.gather()`.  It wraps the passed coroutines
     with a :class:`Supervisor` to ensure the termination of them when returned.
@@ -119,13 +115,11 @@ async def gather_safe(
 
     .. versionadded:: 2.0
     """
-    tasks = []
-    t: asyncio.Task[T]
+    tasks: list[Awaitable[T]] = []
     async with Supervisor(context=context) as supervisor:
         for coro in coros:
             t = supervisor.create_task(coro)
             tasks.append(t)
-        # To ensure safety, the Python version must be 3.7 or higher.
         return await asyncio.gather(*tasks, return_exceptions=True)
 
 
@@ -134,7 +128,7 @@ async def race(
     *,
     continue_on_error: bool = False,
     context: Optional[Context] = None,
-) -> Tuple[T, Sequence[Exception]]:
+) -> tuple[T, Sequence[Exception]]:
     """
     Returns the first result and cancelling all remaining coroutines safely.
     Passing an empty iterable of coroutines is not allowed.